Split numeric type set definitions to new module
This resolves circular import referenecs
