+++ pyomo/common/dependencies.py
@@ -587,16 +587,16 @@ def _finalize_matplotlib(module, available):
 def _finalize_numpy(np, available):
     if not available:
         return
-    import pyomo.core.expr.numvalue as _numvalue
-    _numvalue.RegisterBooleanType(np.bool_)
+    import pyomo.core.expr.numeric_types as _numerictypes
+    _numerictypes.RegisterBooleanType(np.bool_)
     for t in (np.int_, np.intc, np.intp,
               np.int8, np.int16, np.int32, np.int64,
               np.uint8, np.uint16, np.uint32, np.uint64):
-        _numvalue.RegisterIntegerType(t)
-        _numvalue.RegisterBooleanType(t)
+        _numerictypes.RegisterIntegerType(t)
+        _numerictypes.RegisterBooleanType(t)
     for t in (np.float_, np.float16, np.float32, np.float64, np.ndarray):
-        _numvalue.RegisterNumericType(t)
-        _numvalue.RegisterBooleanType(t)
+        _numerictypes.RegisterNumericType(t)
+        _numerictypes.RegisterBooleanType(t)
 
 
 yaml, yaml_available = attempt_import(
